refactor: remove filteredAndSorted function and refactoring Category
