+++ client/src/utils/index.js
@@ -1,2 +1,3 @@
-export { default as filteredAndSortedProducts } from './filterAndSort';
+export { default as filteredProducts } from './filteredProducts';
+export { default as sortProducts } from './sortProducts';
 export { getDecodeSearch } from './location';
